Seperated controller. Moved from cache to cacheable function.
