+++ cacheables/__init__.py
@@ -3,3 +3,4 @@ from cacheables.decorator import *
 from cacheables.caches import *
 from cacheables.exceptions import *
 from cacheables.serializers import *
+from cacheables.controller import enable_all_caches, disable_all_caches
