+++ cacheables/caches/__init__.py
@@ -1,2 +1,2 @@
-from .cache import Cache, enable_all_caches, disable_all_caches
+from .cache import Cache
 from .disk import DiskCache
